Correct some things
