+++ proxy/mempool/__init__.py
@@ -1,7 +1,6 @@
 from .pickable_data_server import PickableDataServer
-from .mempool_client import MemPoolClient
+from .pickable_data_client import PickableDataClient
 from .mempool_service import MemPoolService
-from .mempool_client import MemPoolClient
 from .mem_pool import MemPool
 
 MEMPOOL_SERVICE_PORT = MemPoolService.MEMPOOL_SERVICE_PORT
