+++ proxy/neon_rpc_api_model/neon_rpc_api_model.py
@@ -21,7 +21,7 @@ from ..environment import SOLANA_URL, PP_SOLANA_URL, PYTH_MAPPING_ACCOUNT, NEON_
 from ..memdb.memdb import MemDB
 from ..common_neon.gas_price_calculator import GasPriceCalculator
 from ..statistics_exporter.proxy_metrics_interface import StatisticsExporter
-from ..mempool import MemPoolClient, MEMPOOL_SERVICE_HOST, MEMPOOL_SERVICE_PORT
+from ..mempool import PickableDataClient, MEMPOOL_SERVICE_HOST, MEMPOOL_SERVICE_PORT
 
 from .transaction_sender import NeonTxSender
 from .operator_resource_list import OperatorResourceList
@@ -48,7 +48,7 @@ class NeonRpcApiModel:
         self._solana = SolanaInteractor(SOLANA_URL)
         self._db = MemDB(self._solana)
         self._stat_exporter: Optional[StatisticsExporter] = None
-        self._mempool_client = MemPoolClient(MEMPOOL_SERVICE_HOST, MEMPOOL_SERVICE_PORT)
+        self._mempool_client = PickableDataClient(MEMPOOL_SERVICE_HOST, MEMPOOL_SERVICE_PORT)
 
         if PP_SOLANA_URL == SOLANA_URL:
             self.gas_price_calculator = GasPriceCalculator(self._solana, PYTH_MAPPING_ACCOUNT)
@@ -452,7 +452,7 @@ class NeonRpcApiModel:
 
             self._stat_tx_success()
             neon_tx_data = NeonTxData(tx_signed=rawTrx)
-            self._mempool_client.on_eth_send_raw_transaction(neon_tx_data)
+            self._mempool_client.send_data(neon_tx_data)
             return eth_signature
 
         except PendingTxError as err:
